organize nvim

+++ cmd/nvim/lazy.sh
@@ -1,16 +1,11 @@
-#!/usr/bin/env bash
+#!/bin/bash
 
-set -e
+set -euo pipefail
 
-# ------------------------------------------------------------------------------
-
-mkdir -p "$NVIM_CONFIG_DIR"
-mkdir -p "$S7ISOL_NVIM_CONFIG_DIR"
-
-# ------------------------------------------------------------------------------
+if [[ -z "${NVIM_CONFIG_DIR:-}" ]]; then
+  exit 1
+fi
 
 rm -rf $NVIM_CONFIG_DIR
 git clone https://github.com/LazyVim/starter $NVIM_CONFIG_DIR
 rm -rf $NVIM_CONFIG_DIR/.git
-
-# ------------------------------------------------------------------------------
